# Conflicts:
#	uctp-modules/uctp-bpm/pom.xml
#	uctp-ui/src/views/workbench/components/reverse.vue

+++ uctp-modules/uctp-business/src/main/java/com/newtouch/uctp/module/business/controller/app/account/cash/AccountCashController.java
@@ -62,7 +62,7 @@ public class AccountCashController {
 
     @PostMapping("/difference")
     @Operation(summary = "待回填保证金")
-    public CommonResult<Integer> difference(@Validated(ValidatedGroup.Difference.class) @RequestBody TransactionRecordReqVO transactionRecordReqVO) {
+    public CommonResult<Long> difference(@Validated(ValidatedGroup.Difference.class) @RequestBody TransactionRecordReqVO transactionRecordReqVO) {
         return CommonResult.success(accountCashService.difference(transactionRecordReqVO));
     }
